NIC Build Infra (#6215)
Add module.mk to all build targets
Add support for Makefile 2nd pass for post-generation targets.
add support for non-local files in module.mk
run.py changes to change paths from bazel-bin to build/<ARCH>
Move all Jobd targets to nic/mkdefs/jobd_targets.mk
Move protosync to a common codesync job across workspace.
Split mkinfra and mkdefs
Add pipeline support to module.mk
Add ARCH support to module.mk
Remove fPIE for arm. It causes issue with thread-local-storage
initialization on arm.
Move agent tests into 1 target.

+++ nic/hal/plugins/cfg/nw/session.cc
@@ -16,7 +16,7 @@
 #include "nic/include/pd_api.hpp"
 #include "nic/hal/plugins/cfg/nw/interface_api.hpp"
 #include "nic/hal/plugins/cfg/aclqos/qos.hpp"
-#include "sdk/timestamp.hpp"
+#include "nic/sdk/include/sdk/timestamp.hpp"
 #include "nic/include/fte.hpp"
 #include "nic/hal/plugins/sfw/cfg/nwsec_group.hpp"
 
@@ -2264,7 +2264,7 @@ session_set_tcp_state (session_t *session, hal::flow_role_t role,
     if (flow)
         flow->state = tcp_state;
 
-    HAL_TRACE_DEBUG("Updated tcp state to {}", tcp_state);
+    HAL_TRACE_DEBUG("Updated tcp state to {}", (uint32_t)tcp_state);
 }
 
 bool
